Fix HX711 AVDD level
The HX711's analog output (AVDD) should be set below its supply voltage
(the battery voltage) so that the chip's feedback loop can create a
stable excitation voltage that is resilient to load transients. The
R1/R2 resistor ladder was set such that AVDD would always rail to the
supply voltage. Set it to ~2.7-2.8V instead.

Note: it's possible that this doesn't actually matter since the ADC
makes ratiometric measurements (the reference is the same as the supply
voltage) and may actually hurt SNR.