Inject sampling frequency from binary
The dongle development setup is configured with a 10Hz sampling
frequency, whereas everything else uses 80Hz. Allow this parameter to be
injected into the weight task rather than hard-coding it.

The sampling frequency is primarily used for configuring the taring
process. Accidentally using the default 80Hz value with the dongle
results in very slow startup.