Split up gatt module into a ble module tree
The `gatt` module had grown to be embarrassingly large and hard to
follow. Split it up into multiple modules under the `ble` module tree,
which should hopefully make it a little more readable. This greatly
reduces the number of types and functions that need to be `pub` and even
`pub(crate)`, which I'm taking to be a good sign.